+++ src/wadogo/scale/linear.clj
@@ -1,6 +1,6 @@
 (ns wadogo.scale.linear
   (:require [wadogo.common :refer [scale ->ScaleType strip-keys merge-params]]
-            [wadogo.utils :refer [make-norm]]))
+            [wadogo.utils :refer [make-norm ->extent]]))
 
 (set! *warn-on-reflection* true)
 (set! *unchecked-math* :warn-on-boxed)
@@ -9,9 +9,9 @@
   ([_] (scale :linear {}))
   ([s params]
    (let [params (merge-params s params)
-         [dstart dend] (:domain params)
-         [rstart rend] (:range params)]
-     (->ScaleType :linear (:domain params) (:range params) (:ticks params) (:formatter params)
+         [dstart dend] (->extent (:domain params))
+         [rstart rend] (->extent (:range params))]
+     (->ScaleType :linear [dstart dend] [rstart rend] (:ticks params) (:formatter params)
                   (make-norm dstart dend rstart rend)
                   (make-norm rstart rend dstart dend)
                   (strip-keys params)))))