+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## [1.0.0-SNAPSHOT]
+
+library revision and `clerk` notebook documentation
+
+### Changed
+
+* `bands` scale returns a point by default (was: band info)
+* numerical scales can accept data (extent is calculated then)
+
 ## [0.1.0-SNAPSHOT]
 
 initial release